Add files via upload
Hello everyone,

I have reproduced the error and came up with the conclusion that LinkedIn most probably blocked the User-Agent signature because after I have changed it to IE without touching anything else the software worked properly. I suspect that some of you if not all do not change the browser signature and this leads to high traffic to this particular agent which is suspicious. In the current updated version, I have added an option so you can select from the existing or add your browser agent signature so everyone can be unique. Should you wonder how/or where to find the user agent., please check and select some from the list below:
[https://udger.com/resources/ua-list ](url)

It is recommended to change the User-Agent and use VPN/Proxy regularly to avoid your IP get blocked. You may get "Too many requests" error as Linked has a limited amount of requests per account (per account! not IP) a day. If you get such an error it will be restored in the next 12-24 hours.

From what I have examined during the software testing, you can download around 8-10GB a day before getting that error if it ever occurs. In future updates as mentioned, I will add an option to delay the download time to reduce the chance of happening.

All the best,
r00tme