Lynda Courses Dumper
Sould you need a complete lynda/linkedin courses list to choose a particular course from, use that small script to have the list up to date.

The script create a text file with all available courses after all categories and possible pages are checked. If you interupt the program while it is still running a file won't be created since it's was still in the memory which is vlolatile.