Add support for downloading the Chrome driver that matches the installed instance of Chrome
By importing the ChromeDriverManager and using it to select and download the Chrome Driver to match what is installed on the system, we resolve the reported issue of people failing to run the script because the ChromeDriver only supports Chrome version 80.

I also implemented some format cleanup on Loader.py and a few miscellaneous PEP-8 standards.